fix warning
SVN revision: 55472

+++ legacy/evil/src/lib/evil_unistd.c
@@ -72,8 +72,8 @@ evil_gettimeofday(struct timeval *tp, void *tzp __UNUSED__)
 
    QueryPerformanceCounter(&count);
    diff = count.QuadPart - _evil_time_count;
-   tp->tv_sec = _evil_time_second + diff / _evil_time_freq;
-   tp->tv_usec = ((diff % _evil_time_freq) * 1000000ll) / _evil_time_freq;
+   tp->tv_sec = _evil_time_second + (long)(diff / _evil_time_freq);
+   tp->tv_usec = (long)(((diff % _evil_time_freq) * 1000000ll) / _evil_time_freq);
 
    return 1;
 }
